Fix race condition checking the early warning flag.
Obvious in hindsight. We would read the "triggered" flag as false and then
the IRQ would fire and we'd set to true and then the line that set the
flag to false would run. So we'd never see the true. This defers the
setting to false when we clear it. The only reason this wasn't how it was
before was because I thought there may be situations where we checked in
a loop and didn't want to clear it. This never happens.